requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
blis==0.7.4
catalogue==2.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cymem==2.0.5
discord.py==1.6.0
idna==2.10
Jinja2==2.11.3
MarkupSafe==1.1.1
multidict==5.1.0
murmurhash==1.0.5
numpy==1.20.1
packaging==20.9
pathy==0.4.0
preshed==3.0.5
pydantic==1.7.3
pyparsing==2.4.7
requests==2.25.1
smart-open==3.0.0
spacy==3.0.5
spacy-legacy==3.0.1
srsly==2.4.0
thinc==8.0.2
tqdm==4.59.0
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.3
wasabi==0.8.2
yarl==1.6.3
